+++ plugins/main/public/components/overview/it-hygiene/networks/inventories/networks/dashboard.ts
@@ -1,11 +1,10 @@
 import { buildDashboardKPIPanels } from '../../../common/create-dashboard-panels-kpis';
-import { generateVisualization } from '../../../common/create-new-visualization';
 import { HEIGHT, STYLE } from '../../../common/saved-vis/constants';
 import {
   createIndexPatternReferences,
   createSearchSource,
 } from '../../../common/saved-vis/create-saved-vis-data';
-import { getVisStateDonutByField } from '../../../common/saved-vis/generators';
+import { getVisStateHorizontalBarSplitSeries } from '../../../../../../services/visualizations';
 import { SavedVis } from '../../../common/types';
 import {
   getVisStateDHCPEnabledInterfacesMetric,
@@ -62,11 +61,20 @@ const getVisStateUniqueNetworkIPsMetric = (
 
 export const getOverviewNetworksNetworksTab = (indexPatternId: string) => {
   return buildDashboardKPIPanels([
-    getVisStateDonutByField(
+    getVisStateHorizontalBarSplitSeries(
       indexPatternId,
       'network.type',
-      'Protocols',
+      'Network type',
       'it-hygiene-networks',
+      {
+        fieldSize: 4,
+        otherBucket: 'Others',
+        metricCustomLabel: 'Network type count',
+        valueAxesTitleText: 'Network type count',
+        seriesLabel: 'Type',
+        seriesMode: 'stacked',
+        fieldCustomLabel: 'Type',
+      },
     ),
     getVisStateDHCPEnabledInterfacesMetric(indexPatternId),
     getVisStateUniqueNetworkIPsMetric(indexPatternId),