+++ plugins/main/public/components/overview/it-hygiene/processes/inventories/processes/dashboard.ts
@@ -4,10 +4,7 @@ import {
   createIndexPatternReferences,
   createSearchSource,
 } from '../../../common/saved-vis/create-saved-vis-data';
-import {
-  getVisStateHistrogramBy,
-  getVisStateHorizontalBarByField,
-} from '../../../common/saved-vis/generators';
+import { getVisStateHorizontalBarByField } from '../../../common/saved-vis/generators';
 import { SavedVis } from '../../../common/types';
 
 type ProcessState =
@@ -90,12 +87,12 @@ export const getOverviewProcessesProcessesTab = (indexPatternId: string) => {
       'it-hygiene-processes',
       { customLabel: 'Processes', excludeTerm: '.*wazuh.*' },
     ),
-    getVisStateHistrogramBy(
+    getVisStateHorizontalBarByField(
       indexPatternId,
-      'process.start',
-      'Processes initiation',
+      'agent.name',
+      'Top 5 agents',
       'it-hygiene-processes',
-      'h',
+      { customLabel: 'Processes', excludeTerm: '.*wazuh.*' },
     ),
   ]);
 };